[+][Feat] Now able to have 2 players and more connected in the same room. No motion yet. Refactor coordinates in Float by Double. Issue with the compilation due to failure to load model in another thread than the main.

+++ src/Command/Commands/Player/Move/Move.cpp
@@ -7,7 +7,7 @@
 
 #include "Move.hh"
 
-void Rtype::Command::Player::Move::set_server(std::map<int, std::shared_ptr<Rtype::client_info>> players, int playerID, float x, float y)
+void Rtype::Command::Player::Move::set_server(std::map<int, std::shared_ptr<Rtype::client_info>> players, int playerID, double x, double y)
 {
     _players = players;
     _playerID = playerID;
@@ -15,7 +15,7 @@ void Rtype::Command::Player::Move::set_server(std::map<int, std::shared_ptr<Rtyp
     _y = y;
 }
 
-void Rtype::Command::Player::Move::set_client(udp::endpoint endpoint, float x, float y)
+void Rtype::Command::Player::Move::set_client(udp::endpoint endpoint, double x, double y)
 {
     _endpoint = endpoint;
     _x = x;
