+++ src/Command/Commands/Projectile/Fired/Fired.hh
@@ -20,8 +20,8 @@ namespace Rtype
                     Fired() = default;
                     ~Fired();
 
-                    void set_server(std::map<int, std::shared_ptr<Rtype::client_info>> players, int projectileType, int projectileID, int xOrigin, int yOrigin, int xVector, int yVector);
-                    void set_client(udp::endpoint endpoint, int projectileType, int projectileID, int xVector, int yVector);
+                    void set_server(std::map<int, std::shared_ptr<Rtype::client_info>> players, int projectileType, int projectileID, double xOrigin, double yOrigin, double xVector, double yVector);
+                    void set_client(udp::endpoint endpoint, int projectileType, int projectileID, double xVector, double yVector);
 
                     void execute_client_side();
                     void execute_server_side();
@@ -32,10 +32,10 @@ namespace Rtype
                     std::map<int, std::shared_ptr<Rtype::client_info>> _players;
                     int _projectileType;
                     int _projectileID;
-                    int _xOrigin;
-                    int _yOrigin;
-                    int _xVector;
-                    int _yVector;
+                    double _xOrigin;
+                    double _yOrigin;
+                    double _xVector;
+                    double _yVector;
             };
         }
     }
